+++ packages/xchain-litecoin/src/types/client-types.ts
@@ -1,4 +1,4 @@
-import { FeeOptionKey, Fees } from '@xchainjs/xchain-client'
+import { Address, FeeOptionKey, Fees, Network } from '@xchainjs/xchain-client'
 
 export type FeeRate = number
 export type FeeRates = Record<FeeOptionKey, FeeRate>
@@ -7,3 +7,10 @@ export type FeesWithRates = { rates: FeeRates; fees: Fees }
 
 export type NormalTxParams = { addressTo: string; amount: number; feeRate: number }
 export type VaultTxParams = NormalTxParams & { memo: string }
+
+export type GetChangeParams = {
+  valueOut: number
+  nodeUrl: string
+  network: Network
+  address: Address
+}